issue#2 separate flow
